refactor: 입력값 로또번호로 만들어주는 파싱클래스 패키지명 io에서 parser로 수정
