+++ src/main/java/step3/controller/LottoController.java
@@ -1,6 +1,6 @@
 package step3.controller;
 
-import step3.io.LottoInputParser;
+import step3.parser.LottoInputParser;
 import step3.model.LottoMachine;
 import step3.model.LottoNumber;
 import step3.model.dto.LottoResultDto;
